# optimize combine! for DiscretePartition